pyproject.toml
[project]
name = "intelligent-ticket-router"
version = "0.1.0"
description = "An observable, multi-agent RAG system for intelligent customer support ticket routing using LangChain and Langfuse"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "langfuse>=3.10.1",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "tiktoken>=0.12.0",
]
